!server_info  command script added
The Discord bot will now respond to !server_info commands on any channels of servers that the Bot is in.  Because the bot can be
run on several different DCS servers at once, there is a separate instance of the bot that runs and only listens for this command. 
The bot will then call a bat file, that sets up the python environment, then calls the python script gets the IP addresses, and tests the DCS and SRS ports to see if they are available.